# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChefStudio API",
    "version": "1.0.0",
    "description": "Documentação da API do ChefStudio"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/notifications": {
      "get": {
        "summary": "Obter notificações do usuário",
        "tags": ["Notificações"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Número máximo de notificações a retornar"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Número de notificações a pular"
          },
          {
            "name": "read",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filtrar por notificações lidas/não lidas"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de notificações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "60d21b4667d0d8992e610c85"
                          },
                          "title": {
                            "type": "string",
                            "example": "Campanha ativada"
                          },
                          "message": {
                            "type": "string",
                            "example": "Sua campanha 'Campanha de Verão' foi ativada com sucesso."
                          },
                          "type": {
                            "type": "string",
                            "example": "success"
                          },
                          "read": {
                            "type": "boolean",
                            "example": false
                          },
                          "entityType": {
                            "type": "string",
                            "example": "campaign"
                          },
                          "entityId": {
                            "type": "string",
                            "example": "123456789"
                          },
                          "createdAt": {
                            "type": "string",
                            "example": "2023-06-01T12:00:00Z"
                          }
                        }
                      }
                    },
                    "unreadCount": {
                      "type": "number",
                      "example": 5
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token não fornecido"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao obter notificações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Erro ao obter notificações"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "summary": "Marcar notificação como lida",
        "tags": ["Notificações"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da notificação"
          }
        ],
        "responses": {
          "200": {
            "description": "Notificação marcada como lida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Notificação marcada como lida"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Notificação não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Notificação não encontrada"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token não fornecido"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao marcar notificação como lida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Erro ao marcar notificação como lida"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications/read-all": {
      "patch": {
        "summary": "Marcar todas as notificações como lidas",
        "tags": ["Notificações"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Todas as notificações marcadas como lidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Todas as notificações marcadas como lidas"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token não fornecido"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao marcar todas as notificações como lidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Erro ao marcar todas as notificações como lidas"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}